Lempzev minor fixes and GLORIOUS COMPLETION OF HUFFMAN
James returned the email and I finally understand what went wrong.  It
was because I was treating each file as a text file as opposed to a
binary file.  As a result, main changed, but now it's much more easier
to manage because we're no longer dealing with fstreams but with
vectors!!